+++ playground/framework/react-router.config.ts
@@ -1,3 +1,7 @@
 import type { Config } from "@react-router/dev/config";
 
-export default {} satisfies Config;
+export default {
+  future: {
+    unstable_subResourceIntegrity: true,
+  },
+} satisfies Config;